+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.7 / YYYY-MM-DD
 
+* Rename packages from `io.github.weblegacy.tiles.request.servlet` to `org.apache.tiles.request.jakarta.servlet`
 * Change `license-header.txt` and `NOTICE.txt`
 * delete `bin`-directory
 * Change `.gitignore`